appops: Load the app list asynchronously
The AppList is pretty big and it is loading in the UI thread. This
is noticeable even on high-end chips and there is a definite possibility
that it throws an ANR on lower-end ones.

This patchset adds an asynchronous loader and updates the privacy guard
to use it.

Change-Id: I81f3fb64604af07a351f8cbdfffa7454389e2cee

(cherry picked from commit 02bbdf74a473f3e277b578b67fa96a982bf3082a)

(cherry picked from commit 02bbdf74a473f3e277b578b67fa96a982bf3082a)